requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
braceexpand==0.1.7
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1761745218276/work
certifi==2025.11.12
cffi @ file:///C:/miniconda3/conda-bld/cffi_1760098049787/work
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
diffusers==0.36.0
einops==0.8.1
eval_type_backport==0.3.1
exceptiongroup==1.3.1
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
lightning-utilities==0.15.2
lpips==0.1.4
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.7.0
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy==2.0.2
omegaconf==2.3.0
packaging==25.0
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
platformdirs==4.4.0
propcache==0.4.1
protobuf==6.33.2
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496061107/work
pydantic==2.12.5
pydantic_core==2.41.5
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753077861/work
pytorch-lightning==2.6.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.13.1
sentry-sdk==2.47.0
smmap==5.0.2
sympy @ file:///C:/b/abs_f1v7nwa6vp/croot/sympy_1756713483242/work
tokenizers==0.13.3
torch==2.1.0
torchaudio==2.1.0
torchdiffeq==0.2.5
torchmetrics==1.8.2
torchvision==0.16.0
tqdm==4.67.1
transformers==4.31.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
wandb==0.23.1
webdataset==0.2.100
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746274727/work
yarl==1.22.0
zipp==3.23.0

